requirements.txt
streamlit
langchain-community
pymysql
requests
pypdf
faiss-cpu
langgraph
python-dotenv
nest_asyncio
plotly
